% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_classImpulseDE2Object.R
\name{append_strReport}
\alias{append_strReport}
\title{Append string to strReport in ImpulseDE2Object}
\usage{
append_strReport(obj, s)
}
\arguments{
\item{obj}{(ImpulseDE2Object)  A ImpulseDE2 output object.}

\item{s}{(str) String to be appended.}
}
\value{
(obj)
The ImpulseDE2 object with the str appended in strReport.
}
\description{
Append string to strReport in ImpulseDE2Object.
}
\author{
David Sebastian Fischer
}
