% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_calcImpulse.R
\name{calcImpulse}
\alias{calcImpulse}
\alias{calcImpulse_comp}
\title{Compute value of impulse function given parameters.}
\usage{
calcImpulse(vecTheta, vecTimepoints)
}
\arguments{
\item{vecTheta}{(vector number of parameters) Numerical vector of impulse 
parameters with the order beta, h0, h1, h2, t1, t2.}

\item{vecTimepoints}{(vector number vecTimepoints) 
Observed vecTimepoints, numeric.}
}
\value{
vecY (vec number of vecTimepoints) 
   Model expression values of given gene for time points
}
\description{
Compute value of impulse function given parameters.
}
\seealso{
Called by \code{evalLogLikImpulse},\code{evalLogLikMean}, 
   \code{plotDEGenes}.
}

