% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_classImpulseDE2Object.R
\docType{methods}
\name{$,ImpulseDE2Object-method}
\alias{$,ImpulseDE2Object-method}
\title{List-like accessor methods for ImpulseDE2Object: $}
\usage{
\S4method{$}{ImpulseDE2Object}(x, name)
}
\arguments{
\item{x}{(ImpulseDE2Object) ImpulseDE2 output object.}

\item{name}{(str) Element from x list to be retrieved.}
}
\value{
Target element from x.
}
\description{
$ accessor function for ImpulseDE2Object, relies on [[ ]].
Allow usage of ImpulseDE2 ouput object like a list with
respect to the most relevant output:
dfImpulseDE2Results and vecDEGenes.
List of all available list-object like accessors:
\link{names,ImpulseDE2Object-method},
\link{[[,ImpulseDE2Object,character,missing-method},
\link{$,ImpulseDE2Object-method}.
}

