% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_compareDEMethods.R
\name{compareDEMethods}
\alias{compareDEMethods}
\title{Compare ImpulseDE2 output against other differential expression method}
\usage{
compareDEMethods(matQval, strMethod1 = "ImpulseDE2", strMethod2 = "",
  Q = 10^(-3), Qdelta = 10^(2), matCountDataProc, matTranslationFactors,
  matSizeFactors, dfAnnotationProc, lsImpulseFits, dfImpulseResults,
  strCaseName = "case", strControlName = NULL, strMode = "batch",
  strDataDescriptionFilename = "")
}
\arguments{
\item{matQval:}{matrix with p-values by methods}
}
\description{
The comparison is performed based on the adjusted p-values of differential
expression. This methods allows the user to explore visually how 
ImpulseDE2 output differs from similar methods.
}
\seealso{
Auxillary method not called during ImpulseDE2 running.
Called separately by user.
}

