% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_fitImpulse.R
\name{computeLogLikNull}
\alias{computeLogLikNull}
\title{Fit and compute likelihood of null model for a single gene}
\usage{
computeLogLikNull(vecCounts, scaDispersionEstimate, vecDropoutRate = NULL,
  vecProbNB = NULL, vecNormConst, vecLongitudinalSeries = NULL,
  vecLongitudinalSeriesAssign = NULL, vecboolObserved, vecboolZero = NULL,
  vecboolNotZeroObserved = NULL, strMode)
}
\arguments{
\item{vecLongitudinalSeriesAssign}{(numeric vector number samples) 
Longitudinal sample series assigned to samples. 
NULL if not operating in strMode="longitudinal".}

\item{vecCounts:}{(count vector number of samples) Count data.}

\item{scaDispersionEstimate:}{(scalar) Negative binomial 
dispersion coefficient for given gene.}

\item{vecDropoutRate:}{(probability vector number of samples) 
[Default NULL] Dropout rate/mixing probability of zero inflated 
negative binomial mixturemodel for each gene and cell.}

\item{vecProbNB:}{(probability vector number of samples) 
Probability of observations to come from negative binomial 
component of mixture model.}

\item{vecNormConst:}{(numeric vector number of samples) 
Normalisation constants for each sample.}

\item{vecLongitudinalSeries:}{(string vector number of longitudinal
sample series) Longitudinal sample series.}

\item{vecboolObserved:}{(bool vector number of samples)
Whether sample is observed (not NA).}

\item{vecboolZero:}{(bool vector number of samples)
Whether sample has zero count.}

\item{vecboolNotZeroObserved:}{(bool vector number of samples)
Whether sample is not zero and observed (not NA).}

\item{strMode:}{(str) [Default "batch"] 
{"batch","longitudinal","singlecell"}
Mode of model fitting.}
}
\value{
(list length 3)
   \itemize{
     \item scaMu: (scalar) Null model in batch or
       singlecell mode: One overall mean.
     \item vecMuLongitudinalSeries: (numerical vector
       length number of longitudinal series) Null model
       in longitudinal mode: One mean per longitudinal
       series. Null if mode is not longitudinal.
     \item scaLogLikNull: (scalar) Log likelihood of null
       model.
     }
}
\description{
Fits an impulse model and a mean model to a single gene.
The optimisation method is set within this function (\code{optim}:
BFGS). This method is divided into four parts: (I) Prepare data,
(II) Fit mean model, (III) Fit impulse model, (IV) Process Fits.
Internal function of \code{fitImpulse}.
}
\seealso{
Called by \code{fitImpulse_gene}.
}

