% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_modelFits.R
\name{computeModelFits}
\alias{computeModelFits}
\title{Return a gene x sample and gene x time point (for each condition) matrix with model fits.}
\usage{
computeModelFits(objectImpulseDE2)
}
\arguments{
\item{objectImpulseDE2}{(instance of class ImpulseDE2Object)
ImpulseDE2 output object to create heatmap from.}
}
\value{
\itemize{
\item sample (data.frame)
Data frame with model fit by gene and sample.
\item case (data.frame)
Data frame with model fit by gene and time point in case condition
\item control (data.frame)
Data frame with model fit by gene and time point in control condition
}
}
\description{
Return a gene x sample and gene x time point (for each condition) matrix with model fits.
}
\examples{
lsSimulatedData <- simulateDataSetImpulseDE2(
vecTimePointsA   = rep(seq(1,8),3),
vecTimePointsB   = NULL,
vecBatchesA      = do.call(c, lapply(c("A", "B", "C"), function(x) rep(x, 8))),
vecBatchesB      = NULL,
scaNConst        = 0,
scaNImp          = 50,
scaNLin          = 0,
scaNSig          = 50,
scaMuBatchEffect = 1, 
scaSDBatchEffect = 1)
objectImpulseDE2 <- runImpulseDE2(
matCountData    = lsSimulatedData$matObservedCounts, 
dfAnnotation    = lsSimulatedData$dfAnnotation,
boolCaseCtrl    = FALSE,
vecConfounders  = c("Batch"),
boolIdentifyTransients = TRUE,
scaNProc        = 1 )
modelFits <- computeModelFits(objectImpulseDE2=objectImpulseDE2)
head(modelFits$case)

}
\seealso{
Called seperately by user.
}
\author{
David Sebastian Fischer
}
