% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_computeNormConst.R
\name{computeNormConst}
\alias{computeNormConst}
\title{Compute normalisation constant for each replicate}
\usage{
computeNormConst(matCountDataProcFull, matCountDataProc, vecDispersions,
  scaSmallRun = NULL, dfAnnotationProc, strCaseName, strControlName = NULL,
  strMode = "batch", vecSizeFactorsExternal = NULL)
}
\arguments{
\item{strCaseName}{(str) Name of the case condition in \code{dfAnnotationRedFull}.}

\item{matCountDataProcFull:}{(matrix genes x samples)
Count data: Reduced version of \code{matCountData}. 
For internal use. This is the entire data set.}

\item{matCountDataProc:}{(matrix genes x samples)
Count data: Reduced version of \code{matCountData}. 
For internal use. This is the data set reduced to the
genes which are supposed to be analysed.}

\item{scaSmallRun:}{(integer) [Default NULL] Number of rows
on which ImpulseDE2 is supposed to be run, the full
data set is only used for size factor estimation.}

\item{dfAnnotationProc:}{(Table) Processed annotation table. 
Lists co-variables of samples: 
Sample, Condition, Time (numeric), TimeCateg (str)
(and LongitudinalSeries). For internal use.}

\item{strControlName:}{(str) [Default NULL] Name of the control condition in 
\code{dfAnnotationRedFull}.}

\item{strMode:}{(str) [Default "batch"] 
{"batch","longitudinal","singlecell"}
Mode of model fitting.}

\item{vecSizeFactorsExternal:}{(numeric vector number of cells) 
Model scaling factors for each observation which take
sequencing depth into account (size factors). One size
factor per cell. These are supplied to ImpulseDE, if this variable
is not set, size factors are computed in this function.}
}
\value{
(list {lsMatTranslationFactors,matSizeFactors})
   \itemize{
     \item matTranslationFactors: (numeric matrix genes x samples) 
     Model scaling factors for each observation which take
     longitudinal time series mean within a gene into account 
     (translation factors). Computed based based on all samples.
     \item matSizeFactors: (numeric matrix genes x samples) 
       Model scaling factors for each observation which take
       sequencing depth into account (size factors). One size
       factor per sample - rows of this matrix are equal.
     }
}
\description{
The normalisation constant is the median of the ratio of gene counts versus
the geomtric gene count mean. There is one normalisation constant per 
replicate. An intuitive alternative would be the sequencing depth, the median
ratio is however less sensitive to highly differentially expressed genes
with high counts (ref. DESeq).
The normalisation constants are used to scale the mean of the
negative binomial model inferred during fitting to the sequencing depth 
of the given sample. The normalisation constants therefore replace
normalisation at the count data level, which is not supposed to be done 
in the framework of ImpulseDE2.
There is the option to supply size factors to this function to override
its size factor choice.
}
\seealso{
Called by \code{runImpulseDE2}. 
Calls \code{computeTranslationFactors} and
\code{computeSizeFactors}.
}

