% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_computePval.R
\name{computePval}
\alias{computePval}
\title{Compute p-values for model fit}
\usage{
computePval(matCountDataProc, vecDispersions, dfAnnotationProc, lsImpulseFits,
  strCaseName = NULL, strControlName = NULL, strMode = "batch",
  NPARAM = 6)
}
\arguments{
\item{vecDispersions}{(vector number of genes) Inverse of gene-wise 
negative binomial dispersion coefficients computed by DESeq2.}

\item{lsImpulseFits}{(list length 2 or 6) List of matrices which
contain parameter fits and model values for given time course for the
case condition (and control and combined if control is present).
Each parameter matrix is called parameter_'condition' and has the form
(genes x [beta, h0, h1, h2, t1, t2, logL_H1, converge_H1, mu, logL_H0, 
) where beta to t2 are parameters of the impulse
model, mu is the single parameter of the mean model, logL are
log likelihoods of full (H1) and reduced model (H0) respectively, converge
is convergence status of numerical optimisation of model fitting by
\code{optim} from \code{stats}. Each value matrix is called
value_'condition' and has the form (genes x time points) and contains the
counts predicted by the impulse model at the observed time points.}

\item{strCaseName}{(str) Name of the case condition in \code{dfAnnotationRedFull}.}

\item{NPARAM}{(scalar) [Default 6] Number of parameters of impulse model.}

\item{matCountDataProc:}{(matrix genes x samples)
Count data: Reduced version of \code{matCountData}. 
For internal use.}

\item{dfAnnotationProc:}{(Table) Processed annotation table. 
Lists co-variables of samples: 
Sample, Condition, Time (numeric), TimeCateg (categorial)
(and LongitudinalSeries). For internal use.}

\item{dfDEAnalysis}{(data frame genes x fitting characteristics)}

\item{strControlName:}{(str) [Default NULL] Name of the control condition in 
\code{dfAnnotationRedFull}.}

\item{strMode:}{(str) [Default "batch"] {"batch","longitudinal","singlecell"}
Mode of model fitting.}
}
\value{
dfDEAnalysis (data frame genes x fitting characteristics) 
   Summary of fitting procedure for each gene.
}
\description{
Compute p-value of differential expression based on chi-squared distribution
of deviance of liklihoods and report summary of statistics.
}
\seealso{
Called by \code{runImpulseDE2}.
}

