% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_computeNormConst.R
\name{computeSizeFactors}
\alias{computeSizeFactors}
\title{Compute size factors for a dataset}
\usage{
computeSizeFactors(matCountDataProc, scaSmallRun = NULL, strMode)
}
\arguments{
\item{matCountDataProc:}{(matrix genes x samples)
Count data: Reduced version of \code{matCountData}. 
For internal use.}

\item{scaSmallRun:}{(integer) [Default NULL] Number of rows
on which ImpulseDE2 is supposed to be run, the full
data set is only used for size factor estimation.}

\item{strMode:}{(str) [Default "batch"] 
{"batch","longitudinal","singlecell"}
Mode of model fitting.}
}
\value{
matSizeFactors: (numeric matrix genes x samples) 
   Model scaling factors for each observation which take
   sequencing depth into account (size factors). One size
   factor per sample - rows of this matrix are equal.
}
\description{
This function computes size factors for each sample
in the dataset and expands them to a matrix of the size
of the dataset.
Size factors scale the negative binomial likelihood
model of a gene to the sequencing depth of each sample.
Note that size factors on bulk and single-cell data are 
computed differently: Median ratio of data to geometric mean
for bul data and normalised relative sequencing depth for
single-cell data.
}
\seealso{
Called by \code{computeNormConst}.
}

