% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_computeNormConst.R
\name{computeTranslationFactors}
\alias{computeTranslationFactors}
\title{Compute translation factors for a dataset}
\usage{
computeTranslationFactors(matCountDataProc, matSizeFactors, vecDispersions,
  scaSmallRun = NULL, dfAnnotationProc, strCaseName, strControlName = NULL)
}
\arguments{
\item{strCaseName}{(str) Name of the case condition in \code{dfAnnotationRedFull}.}

\item{matCountDataProc:}{(matrix genes x samples)
Count data: Reduced version of \code{matCountData}. 
For internal use.}

\item{matSizeFactors:}{(numeric matrix genes x samples) 
Model scaling factors for each observation which take
sequencing depth into account (size factors).}

\item{scaSmallRun:}{(integer) [Default NULL] Number of rows
on which ImpulseDE2 is supposed to be run, the full
data set is only used for size factor estimation.}

\item{dfAnnotationProc:}{(Table) Processed annotation table. 
Lists co-variables of samples: 
Sample, Condition, Time (numeric), TimeCateg (str)
(and LongitudinalSeries). For internal use.}

\item{strControlName:}{(str) [Default NULL] Name of the control condition in 
\code{dfAnnotationRedFull}.}
}
\value{
matTranslationFactors: (numeric matrix genes x samples) 
   Model scaling factors for each observation which take
   longitudinal time series mean within a gene into account 
   (translation factors). Computed based based on all samples.
}
\description{
This function computes translation factors for each sample
per gene in the dataset. Note that the translation factors
are computed with respect to the overall mean, and with 
respect to the condition-wise means if control samples
are given. The rescaling of translation factors for case and
control conditions is not necessary for differential 
expression analysis but gives impulse models which directly
represent the data of their conditions, without scaling.
Translation factors scale the negative binomial likelihood
model of a gene to the mean of a longitudinal series of 
samples of this gene. This function is only callde if 
strMode="longitudinal".
}
\seealso{
Called by \code{computeNormConst}.
}

