% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_fitImpulse.R
\name{estimateImpulseParam}
\alias{estimateImpulseParam}
\title{Estimate impulse model parameter initialisations}
\usage{
estimateImpulseParam(vecTimepoints, vecCounts, vecDropoutRate = NULL,
  vecProbNB = NULL, strSCMode = "clustered", scaWindowRadius = NULL,
  vecTimepointAssign, vecNormConst, strMode)
}
\arguments{
\item{vecTimepoints:}{(numeric vector number of timepoints) 
Time-points at which gene was sampled.}

\item{vecCounts:}{(count vector number of samples) Count data.}

\item{vecTimepointAssign:}{(numeric vector number samples) 
Timepoints assigned to samples.}

\item{vecNormConst:}{(numeric vector number of samples) 
Normalisation constants for each sample.}

\item{strMode:}{(str) [Default "batch"] 
{"batch","longitudinal","singlecell"}
Mode of model fitting.}

\item{vecProbNB:}{(probability vector number of samples) 
Probability of observations to come from negative binomial 
component of mixture model.}
}
\value{
vecParamGuessPeak: (numeric vector number of impulse
   model parameters) Impulse model parameter initialisation 
   corresponding to a peak.
}
\description{
The initialisations reflect intuitive parameter choices corresponding
to a peak and to a valley model.
}
\seealso{
Called by \code{fitImpulse_gene}.
}

