% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_fitSigmoid.R
\name{estimateSigmoidParam}
\alias{estimateSigmoidParam}
\title{Compute up and down sigmoid model parameter initialisations
for data of one gene}
\usage{
estimateSigmoidParam(vecCounts, vecTimepoints, vecSizeFactors, lsvecidxBatch)
}
\arguments{
\item{vecCounts}{(numeric vector number of samples)
Read count data.}

\item{vecTimepoints}{(numeric vector length number of samples)
Time coordinates of each sample.}

\item{vecSizeFactors}{(numeric vector number of samples) 
Model scaling factors for each sample which take
sequencing depth into account (size factors).}

\item{lsvecidxBatch}{(list length number of confounding variables)
List of index vectors. 
One vector per confounding variable.
Each vector has one entry per sample with the index batch
within the given confounding variable of the given sample.
Batches are enumerated from 1 to number of batches.}
}
\value{
(list length 2)
\itemize{
\item peak (numeric vector length 6)
\{beta, h0, h1, t\}
Up model initialisations of sigmoidal model parameters.
\item valley (numeric vector length 6)
\{beta, h0, h1, t\}
Down model initialisations of sigmoidal model parameters.
}
}
\description{
[Model fitting function hierarchy: helper to level 2 out of 3]
This is a fitting helper function which computes parameter intialisations
and does not wrap or execute numerical optimisation.
The up model models a sigmoidal expression increase over time,
the down model a sigmoidal decrease over time.
}
\seealso{
Called by \link{fitSigmoidGene}.
}
\author{
David Sebastian Fischer
}
