% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_evalImpulse.R
\name{evalImpulse}
\alias{evalImpulse}
\title{Compute value of impulse function given parameters.}
\usage{
evalImpulse(vecImpulseParam, vecTimepoints)
}
\arguments{
\item{vecImpulseParam}{(numeric vector number of impulse model parameters)
\{beta, h0, h1, h2, t1, t2\}
Vector of impulse model parameters.}

\item{vecTimepoints}{(numeric vector length number of time points) 
Time points to be evaluated.}
}
\value{
vecImpulseValue (vec number of vecTimepoints) 
Model values for given time points.
}
\description{
Compute value of impulse function given parameters.
Enforces lower bound on value of function to avoid numerical
errors during model fitting.
}
\seealso{
Compiled version: \link{evalImpulse_comp}
}
\author{
David Sebastian Fischer
}
