% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_evalImpulse.R
\name{evalImpulse_comp}
\alias{evalImpulse_comp}
\title{Compiled function: evalImpulse}
\usage{
evalImpulse_comp(vecImpulseParam, vecTimepoints)
}
\arguments{
\item{vecImpulseParam}{(numeric vector number of impulse model parameters)
\{beta, h0, h1, h2, t1, t2\}
Vector of impulse model parameters.}

\item{vecTimepoints}{(numeric vector length number of time points) 
Time points to be evaluated.}
}
\value{
vecImpulseValue (vec number of vecTimepoints) 
Model values for given time points.
}
\description{
Pre-compile heavily used functions.
Refer to \link{evalImpulse}.
}
\author{
David Sebastian Fischer
}
