% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_CostFunctionsFit.R
\name{evalLogLikImpulseBatch}
\alias{evalLogLikImpulseBatch}
\alias{evalLogLikImpulseBatch_comp}
\title{Cost function impulse model fit - Batch mode}
\usage{
evalLogLikImpulseBatch(vecTheta, vecX, vecY, scaDispEst, vecNormConst,
  vecindTimepointAssign, vecboolObserved)
}
\arguments{
\item{vecTheta}{(vector number of parameters [6]) 
Impulse model parameters.}

\item{vecX}{(numeric vector number of timepoints) 
Time-points at which gene was sampled.}

\item{vecY}{(count vector samples) 
Observed expression values for given gene.}

\item{vecindTimepointAssign}{(numeric vector number samples) 
Index of time point assigned to sample in list of sorted
time points (vecX).}

\item{scaDispEst:}{(scalar) Dispersion estimate for given gene.}

\item{vecNormConst:}{(numeric vector number of samples) 
Normalisation constants for each sample.}

\item{vecboolObserved:}{(bool vector number of samples)
Stores bool of sample being not NA (observed).}
}
\value{
scaLogLik: (scalar) Value of cost function (likelihood) for given gene.
}
\description{
Log likelihood cost function for impulse model fit based on negative 
binomial model. This cost function is called in the modes
"batch" and "longitudinal".
In analogy to generalised linear models, a log linker
function is used for the count parameters. The inferred negative
binomial model is scaled by the factors in vecNormConst,
which represent size factors (and translation factors),
for evaluation of the likelihood on the data.
}
\seealso{
Called by \code{fitImpulse}:\code{fitImpulse_gene}.
Calls \code{calcImpulse}. \code{evalLogLikImpulseByTC} for
dependent residuals (i.e. time course experiments)
}

