% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_CostFunctionsFit.R
\name{evalLogLikImpulseSC}
\alias{evalLogLikImpulseSC}
\alias{evalLogLikImpulseSC_comp}
\title{Cost function impulse model fit - Single cell mode}
\usage{
evalLogLikImpulseSC(vecTheta, vecX, vecY, scaDispEst, vecDropoutRateEst,
  vecNormConst, vecindTimepointAssign, vecboolNotZeroObserved, vecboolZero,
  scaWindowRadius = NULL)
}
\arguments{
\item{vecTheta}{(vector number of parameters [6]) Impulse model parameters.}

\item{vecX}{(numeric vector number of timepoints) 
Time-points at which gene was sampled.}

\item{vecY}{(count vector samples) Observed expression values for 
given gene.}

\item{vecindTimepointAssign}{(numeric vector number samples) 
Index of time point assigned to sample in list of sorted
time points (vecX).}

\item{scaDispEst:}{(scalar) Negative binomial dispersion 
parameter for given gene.}

\item{vecDropoutRateEst:}{(probability vector number of samples) 
Dropout rate estimate for each cell for given gene.}

\item{vecNormConst:}{(numeric vector number of samples) 
Normalisation constants for each sample.}

\item{vecboolNotZeroObserved:}{(bool vector number of samples)
Whether sample is not zero and observed (not NA).}

\item{vecboolZero:}{(bool vector number of samples)
Whether sample has zero count.}
}
\value{
scaLogLik: (scalar) Value of cost function (likelihood) for given gene.
}
\description{
Log likelihood cost function for impulse model fit based on zero inflated  
negative binomial mode. This cost function is appropriate
for sequencing data with high drop out rate, commonly observed in single
cell data (e.g. scRNA-seq).
In analogy to generalised linear models, a log linker
function is used for the count parameters. The impulse model values
(negative binomial mean parameters) are normalised by vecNormConst
for evaluation of the likelihood on the data.
}
\seealso{
Called by \code{fitImpulse}::\code{fitImpulse_matrix}::
\code{fitImpulse_gene}::\code{optimiseImpulseModelFit}.
Calls \code{calcImpulse} and \code{evalLogLikZINB_comp}.
}

