% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_CostFunctionsFit.R
\name{evalLogLikMu_comp}
\alias{evalLogLikMu_comp}
\title{Compiled function: evalLogLikMu}
\usage{
evalLogLikMu_comp(vecTheta, vecCounts, scaDisp, vecSizeFactors, lsvecidxBatch,
  vecboolObserved)
}
\arguments{
\item{vecTheta}{(numeric vector number of parameters to be estimated) 
Constant model parameter and batch correction factor estimates.}

\item{vecCounts}{(numeric vector number of samples)
Read count data.}

\item{scaDisp}{(scalar) Gene-wise 
negative binomial dispersion hyper-parameter.}

\item{vecSizeFactors}{(numeric vector number of samples) 
Model scaling factors for each sample which take
sequencing depth into account (size factors).}

\item{lsvecidxBatch}{(list length number of confounding variables)
List of index vectors. 
One vector per confounding variable.
Each vector has one entry per sample with the index batch
within the given confounding variable of the given sample.
Batches are enumerated from 1 to number of batches.}

\item{vecboolObserved}{(bool vector number of samples)
Whether sample is observed (finite and not NA).}
}
\value{
scaLogLik (scalar) Value of cost function 
(loglikelihood) for given gene.
}
\description{
Pre-compile heavily used functions.
Refer to \link{evalLogLikMu}.
}
\author{
David Sebastian Fischer
}
