% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_CostFunctionsFit.R
\name{evalLogLikNBMean}
\alias{evalLogLikNBMean}
\alias{evalLogLikNBMean_comp}
\title{Cost function for negative binomial mean parameter fit}
\usage{
evalLogLikNBMean(scaTheta, vecCounts, scaDispEst, vecNormConst, vecboolObserved)
}
\arguments{
\item{scaTheta}{(vector number of parameters [6]) 
Negative binomial mean parameter.}

\item{vecCounts}{(count vector samples) 
Observed expression values for given gene.}

\item{scaDispEst:}{(scalar) Dispersion estimate for given gene.}

\item{vecNormConst:}{(numeric vector number of samples) 
Normalisation constants for each sample.}

\item{vecWeights:}{(probability vector number of samples) 
Weights for inference on mixture models.}

\item{vecboolObserved:}{(bool vector number of samples)
Stores bool of sample being not NA (observed).}
}
\value{
scaLogLik: (scalar) Value of cost function (likelihood) for given gene.
}
\description{
Log likelihood cost function for numerical optimisation of mean model fit 
based on negative binomial model. Note that the closed form solution
of the negative binomial mean parameter only holds if all normalisation
factors are 1.
In analogy to generalised linear models, a log linker
function is used for the mean. The inferred negative
binomial mean model is scaled by the factors in vecNormConst,
which represent size factors (and translation factors),
for evaluation of the likelihood on the data.
}
\seealso{

}

