% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_CostFunctionsFit.R
\name{evalLogLikZINB}
\alias{evalLogLikZINB}
\alias{evalLogLikZINB_comp}
\title{Compute log likelihood of zero-inflated negative binomial model}
\usage{
evalLogLikZINB(vecY, vecMu, scaDispEst, vecDropoutRateEst,
  vecboolNotZeroObserved, vecboolZero)
}
\arguments{
\item{vecY}{(count vector number of amples)
Observed expression values for  given gene.}

\item{vecMu}{(vector number of samples) Negative binomial
mean parameter for each sample.}

\item{scaDispEst:}{(scalar) Negative binomial dispersion 
parameter for given gene.}

\item{vecDropoutRateEst:}{(probability vector number of samples) 
Dropout rate estimate for each cell for given gene.}

\item{vecboolNotZeroObserved:}{(bool vector number of samples)
Whether sample is not zero and observed (not NA).}

\item{vecboolZero:}{(bool vector number of samples)
Whether sample has zero count.}
}
\value{
scaLogLik: (scalar) Value of cost function (likelihood) for given gene.
}
\description{
This liklihood function is appropriate for sequencing data with high drop 
out rate, commonly observed in single cell data (e.g. scRNA-seq).
}
\seealso{
Called by \code{fitImpulse}::\code{fitImpulse_matrix}::
\code{fitImpulse_gene}::\code{computeLogLikNull} and
\code{evalLogLikImpulseSC}.
}

