% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_evalSigmoid.R
\name{evalSigmoid}
\alias{evalSigmoid}
\title{Compute value of sigmoidal model given parameters.}
\usage{
evalSigmoid(vecSigmoidParam, vecTimepoints)
}
\arguments{
\item{vecSigmoidParam}{(numeric vector number of sigmoid model parameters)
\{beta, h0, h1, t1\}
Vector of sigmoidal model parameters.}

\item{vecTimepoints}{(numeric vector length number of time points) 
Time points to be evaluated.}
}
\value{
vecSigmoidValue (numeric vector length of vecTimepoints) 
Model values for given time points.
}
\description{
Compute value of sigmoidal model given parameters.
Enforces lower bound on value of function to avoid numerical
errors during model fitting.
}
\seealso{
Compiled version: \link{evalSigmoid_comp}
}
\author{
David Sebastian Fischer
}
