% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_evalSigmoid.R
\name{evalSigmoid_comp}
\alias{evalSigmoid_comp}
\title{Compiled function: evalSigmoid}
\usage{
evalSigmoid_comp(vecSigmoidParam, vecTimepoints)
}
\arguments{
\item{vecSigmoidParam}{(numeric vector number of sigmoid model parameters)
\{beta, h0, h1, t1\}
Vector of sigmoidal model parameters.}

\item{vecTimepoints}{(numeric vector length number of time points) 
Time points to be evaluated.}
}
\value{
vecSigmoidValue (numeric vector length of vecTimepoints) 
Model values for given time points.
}
\description{
Pre-compile heavily used functions.
Refer to \link{evalSigmoid}.
}
\author{
David Sebastian Fischer
}
