% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_fitImpulse.R
\name{fitConstModel}
\alias{fitConstModel}
\title{Fit a constant model to data of a gene}
\usage{
fitConstModel(vecCounts, scaDisp, vecSizeFactors, lsvecidxBatch, MAXIT = 1000,
  RELTOL = 10^(-8), trace = 0, REPORT = 10)
}
\arguments{
\item{vecCounts}{(numeric vector number of samples)
Read count data.}

\item{scaDisp}{(scalar) Gene-wise 
negative binomial dispersion hyper-parameter.}

\item{vecSizeFactors}{(numeric vector number of samples) 
Model scaling factors for each sample which take
sequencing depth into account (size factors).}

\item{lsvecidxBatch}{(list length number of confounding variables)
List of index vectors. 
One vector per confounding variable.
Each vector has one entry per sample with the index batch
within the given confounding variable of the given sample.
Batches are enumerated from 1 to number of batches.}

\item{MAXIT}{(scalar) [Default 1000] 
Maximum number of BFGS iterations for model fitting with \link{optim}.}

\item{RELTOL}{(scalar) [Default 10^(-8)]
Maximum relative change in loglikelihood to reach convergence in
numerical optimisation by BFGS in \link{optim}.}

\item{trace}{(scalar) [Defaul 0]
Reporting parameter of \link{optim}.}

\item{REPORT}{(scalar) [Default 10]
Reporting parameter of \link{optim}.}
}
\value{
(list) List of constant fit parameters and results.
\itemize{
\item scaMu (scalar) Maximum likelihood estimator of
negative binomial mean parameter.
\item lsvecBatchFactors (list length number of confounders)
List of vectors of scalar batch correction factors for each sample.
These are also maximum likelihood estimators.
NULL if no confounders given.
\item scaDispParam (scalar) Dispersion parameter estimate
used in fitting (hyper-parameter).
\item scaLL (scalar) Loglikelihood of data under maximum likelihood
estimator model.
\item scaConvergence (scalar) 
Convergence status of optim on constant model.
}
}
\description{
[Model fitting function hierarchy: 4 out of 4]
This quarterny fitting wrapper performs constant model fitting:
This function executes numerical optimisaiton and error-handling
thereof.
}
\seealso{
Called by \link{fitConstImpulseGene} to fit constant
model to samples of one condition and one gene.
Calls constant model cost function 
\link{evalLogLikMu} within \link{optim}.
}
\author{
David Sebastian Fischer
}
