% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_fitImpulse.R
\name{fitImpulse}
\alias{fitImpulse}
\title{Fits impulse model to a timecourse dataset}
\usage{
fitImpulse(matCountDataProc, dfAnnotationProc, matTranslationFactors,
  matSizeFactors, vecDispersions, matDropoutRate, matProbNB,
  vecClusterAssignments, strCaseName, strControlName = NULL,
  strMode = "batch", strSCMode = "clustered", scaWindowRadius = NULL,
  nProc = 1, NPARAM = 6)
}
\arguments{
\item{strCaseName}{(str) Name of the case condition in \code{dfAnnotationRedFull}.}

\item{matCountDataProc:}{(matrix genes x samples)
Count data: Reduced version of \code{matCountData}. 
For internal use.}

\item{dfAnnotationProc:}{(Table) Processed annotation table. 
Lists co-variables of samples: 
Sample, Condition, Time (numeric), TimeCateg (str)
(and LongitudinalSeries). For internal use.}

\item{matTranslationFactors:}{(numeric matrix genes x samples) 
Model scaling factors for each observation which take
longitudinal time series mean within a gene into account 
(translation factors). Computed based based on all samples.}

\item{matSizeFactors:}{(numeric matrix genes x samples) 
Model scaling factors for each observation which take
sequencing depth into account (size factors). One size
factor per sample - rows of this matrix are equal.}

\item{vecDispersions:}{(vector number of genes) Gene-wise 
negative binomial dispersion coefficients.}

\item{matDropoutRate:}{(probability matrix genes x samples) Dropout 
rate/mixing probability of zero inflated negative binomial mixture 
model for each gene and cell.}

\item{matProbNB:}{(probability matrix genes x samples) 
Probability of observations to come from negative binomial 
component of mixture model.}

\item{strControlName:}{(str) [Default NULL] Name of the control condition in 
\code{dfAnnotationRedFull}.}

\item{strMode:}{(str) [Default "batch"] 
{"batch","longitudinal","singlecell"}
Mode of model fitting.}

\item{nProc:}{(scalar) [Default 1] Number of processes for parallelisation.}

\item{NPARAM:}{(scalar) [Default 6] Number of parameters of impulse model.}
}
\value{
lsFitResults_all (list length 2 or 6) List of matrices which
   contain parameter fits and model values for given time course for the
   case condition (and control and combined if control is present).
   Each parameter matrix is called parameter_'condition' and has the form
   (genes x [beta, h0, h1, h2, t1, t2, logL_H1, converge_H1, mu, logL_H0, 
   converge_H0]) where beta to t2 are parameters of the impulse
   model, mu is the single parameter of the mean model, logL are
   log likelihoods of full (H1) and reduced model (H0) respectively, converge
   is convergence status of numerical optimisation of model fitting by
   \code{optim} from \code{stats} of either model. Each value matrix is called
   value_'condition' and has the form (genes x time points) and contains the
   counts predicted by the impulse model at the observed time points.
}
\description{
This function processes the input matrix and
coordinates impulse model fitting through
\code{impulse_fit_matrix}.
[Helper \code{fitImpulse_matrix}] Fit impulse model to matrix of genes.
Calls \code{fitImpulse_gene}. \code{fitImpulse_matrix} fits impulse
models to a matrix of samples from one condition.
}
\seealso{
Called by \code{runImpulseDE2}. Calls \code{fitImpulse_matrix}.
}

