% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_fitImpulse.R
\name{fitImpulse_gene}
\alias{fitImpulse_gene}
\title{Fit an impulse model to a single gene}
\usage{
fitImpulse_gene(vecCounts, scaDispersionEstimate, vecDropoutRate = NULL,
  vecProbNB = NULL, vecNormConst, vecTimepointAssign,
  vecLongitudinalSeriesAssign, dfAnnotationProc, strMode = "batch",
  strSCMode = "clustered", scaWindowRadius = NULL, NPARAM = 6,
  MAXIT = 1000)
}
\arguments{
\item{vecLongitudinalSeriesAssign}{(numeric vector number samples) 
Longitudinal series assigned to samples. 
NULL if not operating in strMode="longitudinal".}

\item{vecCounts:}{(count vector number of samples) Count data.}

\item{scaDispersionEstimate:}{(scalar) Inverse of negative binomial 
dispersion coefficients computed by DESeq2 for given gene.}

\item{vecDropoutRate:}{(probability vector number of samples) 
[Default NULL] Dropout rate/mixing probability of zero inflated 
negative binomial mixturemodel for each gene and cell.}

\item{vecProbNB:}{(probability vector number of samples) [Default NULL]
Probability of observations to come from negative binomial 
component of mixture model.}

\item{vecNormConst:}{(numeric vector number of samples) 
Model scaling factors for each observation: Take
sequencing depth and longitudinal time series mean
within a gene into account (size and translation
factors).}

\item{vecTimepointAssign:}{(numeric vector number samples) 
Timepoints assigned to samples.}

\item{dfAnnotationProc:}{(Table) Processed annotation table. 
Lists co-variables of samples: 
Sample, Condition, Time (numeric), TimeCateg (str)
(and LongitudinalSeries). For internal use.}

\item{strMode:}{(str) [Default "batch"] 
{"batch","longitudinal","singlecell"}
Mode of model fitting.}

\item{NPARAM:}{(scalar) [Default 6] Number of impulse model parameters}

\item{MAXIT:}{(scalar) [Default 100] Number of iterations, which are performed 
to fit the impulse model to the clusters.}
}
\value{
vecBestFitSummary: (vector [beta, h0, h1, h2, t1, t2, logL_H1, 
   converge_H1, mu, logL_H0]) Beta to t2 are parameters of the impulse
   model, mu is the single parameter of the mean model, logL are
   log likelihoods of full (H1) and reduced model (H0) respectively, converge
   is convergence status of numerical optimisation of model fitting by
   \code{optim} from \code{stats}.
}
\description{
Fits an impulse model and a mean model to a single gene.
The optimisation method is set within \code{optimiseImpulseModelFit} 
(\code{optim}: BFGS). This function is divided into four parts: 
(I) Prepare data,
(II) Fit mean model, 
(III) Fit impulse model, 
(IV) Process Fits.
The body of this function is broken up into 4 helper functions, 
which are exclusively called by this function.
}
\seealso{
Called by \code{fitImpulse_matrix}. This function
calls \code{computeLogLikNull}, \code{estimateImpulseParam} and
\code{optimiseImpulseModelFit}.
}

