% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_fitImpulse.R
\name{fitImpulse_matrix}
\alias{fitImpulse_matrix}
\title{Fits impulse models to all genes of a dataset}
\usage{
fitImpulse_matrix(matCountDataProcCondition, vecDispersions,
  matDropoutRate = NULL, matProbNB = NULL, matNormConst, vecTimepointAssign,
  vecLongitudinalSeriesAssign, dfAnnotationProc, strCaseName,
  strControlName = NULL, strMode = "batch", strSCMode = "clustered",
  scaWindowRadius = NULL, nProc = 1, NPARAM = 6)
}
\arguments{
\item{matCountDataProc:}{(count matrix genes x samples) Count data.}

\item{vecDispersions:}{(vector number of genes) Gene-wise 
negative binomial dispersion coefficients.}

\item{matDropoutRate:}{(probability matrix genes x samples) Dropout 
rate/mixing probability of zero inflated negative binomial mixture 
model for each gene and cell.}

\item{matProbNB:}{(probability vector genes x samples) 
Probability of observations to come from negative binomial 
component of mixture model.}

\item{matNormConst:}{(numeric matrix genes x samples) 
Model scaling factors for each observation: Take
sequencing depth and longitudinal time series mean
within a gene into account (size and translation
factors).}

\item{vecTimepointAssign:}{(numeric vector number samples) Timepoints 
assigned to samples.}

\item{vecLongitudinalSeriesAssign:}{(numeric vector number samples) Time courses 
assigned to samples. NULL if not operating in strMode="longitudinal".}

\item{dfAnnotationProc:}{(Table) Processed annotation table. 
Lists co-variables of samples: 
Sample, Condition, Time (numeric), TimeCateg (str)
(and LongitudinalSeries). For internal use.}

\item{strCaseName:}{(str) Name of the case condition in \code{dfAnnotationRedFull}.}

\item{strControlName:}{(str) [Default NULL] Name of the control condition in 
\code{dfAnnotationRedFull}.}

\item{strMode:}{(str) [Default "batch"] 
{"batch","longitudinal","singlecell"}
Mode of model fitting.}

\item{nProc:}{(scalar) [Default 3] Number of processes for parallelisation.}

\item{NPARAM:}{(scalar) [Default 6] Number of parameters of impulse model.}
}
\value{
lsFitResults_matrix (list length 2) List of two matrices which
   contain parameter fits and model values for given time course for the
   given condition.
\itemize{
   \item \code{parameter}(genes x [beta, h0, h1, h2, t1, t2, logL_H1, 
       converge_H1, mu, logL_H0, converge_H0]) 
       Beta to t2 are parameters of the impulse model, mu is the single 
       parameter of the mean model, logL are log likelihoods of full 
       (H1) and reduced model (H0) respectively, converge is convergence 
       status of numerical optimisation of model fitting by
       \code{optim} from \code{stats} of either model.
   \item \code{values} (genes x time points) Contains the counts 
       predicted by the impulse model at the observed time points.
}
}
\description{
Fits impulse models to all genes of a dataset. Performs parallelisation.
Internal function of \code{fitImpulse}.
}
\details{
Maximum number of iterations for optimisation is set within this 
   function as \code{MAXIT}. In the case of single condition differential 
   expression over time, this function is called once for the case condition.
   In the case of case and control condition, this function is called three 
   times: data from case, control and combined conditions.
   Calls \code{fitImpulse_gene}.
}
\seealso{
Called by \code{fitImpulse}. Calls \code{fitImpulse_gene}.
}

