% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_CostFunctionsFit.R
\name{fitNBMean}
\alias{fitNBMean}
\title{Fit negative binomial mean parameter}
\usage{
fitNBMean(vecCounts, scaDispEst, vecNormConst)
}
\arguments{
\item{vecCounts}{(count vector samples) 
Observed expression values for given gene.}

\item{scaDispEst:}{(scalar) Dispersion estimate for given gene.}

\item{vecNormConst:}{(numeric vector number of samples) 
Normalisation constants for each sample.}

\item{vecWeights:}{(probability vector number of samples) 
Weights for inference on mixture models.}

\item{vecboolObserved:}{(bool vector number of samples)
Stores bool of sample being not NA (observed).}
}
\value{
scaMu: (scalar) Maximum likelihood estimator of
   negative binomial mean parameter.
}
\description{
Numerical optimisation of negative binomial mean model fit.
Note that the closed form solution of the maximum likelihood 
estimator of the negative binomial mean parameter 
(the weighted average) only holds if all normalisation
factors are 1. Catches numerical errors.
}
\seealso{
Called by \code{computeTranslationFactors()}
and \code{computeLogLikNull()}.
}

