% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_fitSigmoid.R
\name{fitSigmoidGene}
\alias{fitSigmoidGene}
\title{Fit a sigmoidal model to a single gene}
\usage{
fitSigmoidGene(vecCounts, scaDisp, vecSizeFactors, vecTimepointsUnique,
  vecidxTimepoint, lsvecidxBatch, MAXIT = 1000)
}
\arguments{
\item{vecCounts}{(numeric vector number of samples)
Read count data.}

\item{scaDisp}{(scalar) Gene-wise 
negative binomial dispersion hyper-parameter.}

\item{vecSizeFactors}{(numeric vector number of samples) 
Model scaling factors for each sample which take
sequencing depth into account (size factors).}

\item{vecTimepointsUnique}{(numeric vector length number of unique
timepoints) Vector of unique time coordinates observed in this condition.}

\item{vecidxTimepoint}{(idx vector length number of samples)
Index of the time coordinates of each sample (reference is
vecTimepointsUnique).}

\item{lsvecidxBatch}{(idx list length number of confounding variables)
List of vectors. 
One vector per confounding variable.
Each vector has one entry per sample with the index of the batch ID
within the given confounding variable of the given sample. Reference
is the list of unique batch ids for each confounding variable.}

\item{MAXIT}{(scalar) [Default 1000] 
Maximum number of BFGS iterations for model fitting with \link{optim}.}
}
\value{
(list) 
List of sigmoidal fit parameters and results.
\itemize{
\item vecSigmoidParam (numeric vector length 4)
\{beta, h0, h1, t\}
Maximum likelihood estimators of sigmoidal model parameters.
\item vecSigmoidValue (numeric vector length number of time points)
Values of sigmoid model fit at time points used for fit.
\item lsvecBatchFactors (list length number of confounders)
List of vectors of scalar batch correction factors for each sample.
These are also maximum likelihood estimators.
NULL if no confounders given.
\item scaDispParam (scalar) Dispersion parameter estimate
used in fitting (hyper-parameter).
\item scaLL (scalar) Loglikelihood of data under maximum likelihood
estimator model.
\item scaConvergence (scalar) 
Convergence status of optim on sigmoidal model.
}
}
\description{
[Model fitting function hierarchy: 2 out of 3]
This secondary fitting wrapper calls the optimisation wrappers
for the individual fitting operations to be performed on the 
observations of this gene.
Structure of this function:
\itemize{
\item Fit sigmoidal model
\itemize{
\item Initialise sigmoidal model parameters (up and down)
\item Fit sigmoidal model (up initialisation)
\item Fit sigmoidal model (down initialisation)
}
\item Select best sigmoidal model fit from initialisations,
}
}
\seealso{
Called by \link{fitSigmoidModels} to fit
sigmoidal model to samples of one condition and one gene.
Calls sigmoidal parameter initialisation function
\link{estimateSigmoidParam} and 
optimisation wrapper \link{fitSigmoidModel}.
}
\author{
David Sebastian Fischer
}
