% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_classImpulseDE2Object.R
\name{get_accessors}
\alias{get_accessors}
\alias{get_matCountDataProc}
\alias{get_dfAnnotationProc}
\alias{get_vecSizeFactors}
\alias{get_vecDispersions}
\alias{get_boolCaseCtrl}
\alias{get_vecConfounders}
\alias{get_scaNProc}
\alias{get_scaQThres}
\alias{get_strReport}
\alias{get_lsModelFits}
\alias{get_matCountDataProc}
\alias{get_dfAnnotationProc}
\alias{get_vecAllIDs}
\alias{get_vecSizeFactors}
\alias{get_vecDispersions}
\alias{get_boolCaseCtrl}
\alias{get_vecConfounders}
\alias{get_scaNProc}
\alias{get_scaQThres}
\alias{get_strReport}
\title{ImpulseDE2Object "get" accession functions}
\usage{
get_lsModelFits(obj)

get_matCountDataProc(obj)

get_dfAnnotationProc(obj)

get_vecAllIDs(obj)

get_vecSizeFactors(obj)

get_vecDispersions(obj)

get_boolCaseCtrl(obj)

get_vecConfounders(obj)

get_scaNProc(obj)

get_scaQThres(obj)

get_strReport(obj)
}
\arguments{
\item{obj}{(ImpulseDE2Object)  A ImpulseDE2 output object.}
}
\value{
The internal data object specified by the function.
}
\description{
Get internal data of ImpulseDE2 output object.
}
\examples{
   
lsSimulatedData <- simulateDataSetImpulseDE2(
vecTimePointsA   = rep(seq(1,8),3),
vecTimePointsB   = NULL,
vecBatchesA      = NULL,
vecBatchesB      = NULL,
scaNConst        = 30,
scaNImp          = 10,
scaNLin          = 10,
scaNSig          = 10)
objectImpulseDE2 <- runImpulseDE2(
matCountData    = lsSimulatedData$matObservedCounts, 
dfAnnotation    = lsSimulatedData$dfAnnotation,
boolCaseCtrl    = FALSE,
vecConfounders  = NULL,
scaNProc        = 1 )
# Extract hidden auxillary result and processed input objects.
lsModelFits <- get_lsModelFits(objectImpulseDE2)
matCountDataProc <- get_matCountDataProc(objectImpulseDE2)
dfAnnotationProc <- get_dfAnnotationProc(objectImpulseDE2)
vecAllIDs <- get_vecAllIDs(objectImpulseDE2)
vecSizeFactors <- get_vecSizeFactors(objectImpulseDE2)
vecDispersions <- get_vecDispersions(objectImpulseDE2)
boolCaseCtrl <- get_boolCaseCtrl(objectImpulseDE2)
vecConfounders <- get_vecConfounders(objectImpulseDE2)
scaNProc <- get_scaNProc(objectImpulseDE2)
scaQThres <- get_scaQThres(objectImpulseDE2)
strReport <- get_strReport(objectImpulseDE2)

}
\author{
David Sebastian Fischer
}
