% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_classImpulseDE2Object.R
\docType{class}
\name{ImpulseDE2Object-class}
\alias{ImpulseDE2Object-class}
\title{Container class for ImpulseDE2 output}
\description{
ImpulseDE2 output and intermediate results such as model fits.
}
\section{Slots}{

\describe{
\item{\code{dfDEAnalysis}}{(data frame samples x reported characteristics) 
Summary of fitting procedure and 
differential expression results for each gene.
\itemize{
\item Gene: Gene ID.
\item p: P-value for differential expression.
\item padj: Benjamini-Hochberg false-discovery rate corrected p-value
for differential expression analysis.
\item loglik_full: Loglikelihood of full model.
\item loglik_red: Loglikelihood of reduced model.
\item df_full: Degrees of freedom of full model.
\item df_red: Degrees of freedom of reduced model
\item mean: Inferred mean parameter of constant model of first batch.
From combined samples in case-ctrl. 
\item allZero (bool) Whether there were no observed 
non-zero observations of this gene.
If TRUE, fitting and DE analsysis were skipped and entry is NA.
}
Entries only present in case-only DE analysis:
\itemize{
\item converge_impulse: Convergence status of optim for 
impulse model fit (full model).
\item converge_const: Convergence status of optim for 
constant model fit (reduced model).
}
Entries only present in case-control DE analysis:
\itemize{
\item converge_combined: Convergence status of optim for 
impulse model fit to case and control samples combined (reduced model).
\item converge_case: Convergence status of optim for 
impulse model fit to samples of case condition (full model 1/2).
\item converge_control: Convergence status of optim for 
impulse model fit to samples of control condition (full model 2/2).
}
Entries only present if boolIdentifyTransients is TRUE:
\itemize{
\item converge_sigmoid: Convergence status of optim for 
sigmoid model fit to samples of case condition.
\item impulseTOsigmoid_p: P-value of loglikelihood ratio test
impulse model fit versus sigmoidal model on samples of case condition.
\item impulseTOsigmoid_padj: Benjamini-Hochberg 
false-discovery rate corrected p-value of loglikelihood ratio test
impulse model fit versus sigmoid model on samples of case condition.
\item sigmoidTOconst_p: P-value of loglikelihood ratio test
sigmoidal model fit versus constant model on samples of case condition.
\item sigmoidTOconst_padj: Benjamini-Hochberg 
false-discovery rate corrected p-value of loglikelihood ratio test
sigmoidal model fit versus constant model on samples of case condition.
\item isTransient (bool) Whether gene is transiently
activated or deactivated and differentially expressed.
\item isMonotonous (bool) Whether gene is not transiently
activated or deactivated and differentially expressed. This scenario
corresponds to a montonous expression level increase or decrease.
}}

\item{\code{vecDEGenes}}{(list number of genes) Genes IDs identified
as differentially expressed by ImpulseDE2 at threshold \code{scaQThres}.}

\item{\code{lsModelFits}}{(list length number of conditions fit (1 or 3))
{'case'} or {'case', 'control', 'combined'}
One model fitting object for each condition:
In case-only DE analysis, only the condition {'case'} is fit.
In case-control DE analysis, the conditions 
{'case', 'control','combined} are fit.
Each condition entry is a list of model fits for each gene.
Each gene entry is a list of model fits to the individual models:
Impulse model and constant model (if boolFitConst is TRUE).
At this level, the sigmoid model fit can be added later.
Each model fit per gene is a list of fitting parameters and results.
\itemize{
\item IdxGroups (list length number of conditions)
Samples grouped by time points and by batches and time point vectors. 
Sample groups are stored in the form of index vectors in which
samples of the same time point or batch have the same index.
\itemize{
\item Condition ID (list length 3)
List of index vectors and time points.
One entry of this format for each condition.
\itemize{
\item vecTimepointsUnique (numeric vector length number of unique
timepoints) Vector of unique time coordinates observed in this condition.
\item vecidxTimepoint (idx vector length number of samples)
Index of the time coordinates of each sample (reference is
vecTimepointsUnique).
\item lsvecBatchUnique (list number of confounders)
List of string vectors. One vector per confounder: vector of unique batches
in this confounder.
\item lsvecidxBatches (idx list length number of confounding variables)
List of index vectors. 
One vector per confounding variable.
Each vector has one entry per sample with the index of the batch ID
within the given confounding variable of the given sample. Reference
is the list of unique batch ids for each confounding variable.
}
}
\item Condition ID (list length number of genes)
List of fits for each gene to the samples of this condition.
One entry of this format for all conditions fit.
\itemize{
\item Gene ID (list length 2)
Impulse and constant model fit to gene observations.
One entry of this format for all gene IDs.
\itemize{
\item lsImpulseFit (list) List of impulse fit parameters and results.
\itemize{
\item vecImpulseParam (numeric vector length 6)
{beta, h0, h1, h2, t1, t2}
Maximum likelihood estimators of impulse model parameters.
\item vecImpulseValue (numeric vector length number of time points)
Values of impulse model fit at time points used for fit.
\item lsvecBatchFactors (list length number of confounders)
List of vectors of scalar batch correction factors for each sample.
These are also maximum likelihood estimators.
NULL if no confounders given.
\item scaDispParam (scalar) Dispersion parameter estimate
used in fitting (hyper-parameter).
\item scaLL (scalar) Loglikelihood of data under maximum likelihood
estimator model.
\item scaConvergence (scalar) 
Convergence status of optim on impulse model.
}
\item lsConstFit (list) List of constant fit parameters and results.
\itemize{
\item scaMu (scalar) Maximum likelihood estimator of
negative binomial mean parameter.
\item lsvecBatchFactors (list length number of confounders)
List of vectors of scalar batch correction factors for each sample.
These are also maximum likelihood estimators.
NULL if no confounders given.
\item scaDispParam (scalar) Dispersion parameter estimate
used in fitting (hyper-parameter).
\item scaLL (scalar) Loglikelihood of data under maximum likelihood
estimator model.
\item scaConvergence (scalar) 
Convergence status of optim on constant model.
}
\item ls SigmoidFit (list) List of sigmoidal fit parameters and results.
NULL if boolIdentifyTransients is FALSE.
\itemize{
\item vecSigmoidParam (numeric vector length 4)
{beta, h0, h1, t}
Maximum likelihood estimators of sigmoidal model parameters.
\item vecSigmoidValue (numeric vector length number of time points)
Values of sigmoid model fit at time points used for fit.
\item lsvecBatchFactors (list length number of confounders)
List of vectors of scalar batch correction factors for each sample.
These are also maximum likelihood estimators.
NULL if no confounders given.
\item scaDispParam (scalar) Dispersion parameter estimate
used in fitting (hyper-parameter).
\item scaLL (scalar) Loglikelihood of data under maximum likelihood
estimator model.
\item scaConvergence (scalar) 
Convergence status of optim on sigmoidal model.
}
}
}
}}

\item{\code{matCountDataProc}}{(matrix genes x samples) [Default NULL] 
Read count data, unobserved entries are NA. Processed matrix.}

\item{\code{dfAnnotationProc}}{(data frame samples x covariates) 
{Sample, Condition, Time (numeric), TimeCateg (str)
(and confounding variables if given).}
Annotation table with covariates for each sample.
Processed table.}

\item{\code{vecDispersions}}{(numeric vector number of samples) 
Gene-wise negative binomial dispersion hyper-parameters.}

\item{\code{vecSizeFactors}}{(numeric vector number of samples) 
Model scaling factors for each sample which take
sequencing depth into account (size factors).}

\item{\code{boolCaseCtrl}}{(bool) 
Whether to perform case-control analysis. Does case-only
analysis if FALSE.}

\item{\code{vecConfounders}}{(vector of strings number of confounding variables)
Factors to correct for during batch correction. Have to 
supply dispersion factors if more than one is supplied.
Names refer to columns in dfAnnotation.}

\item{\code{scaNProc}}{(scalar) Number of processes for 
parallelisation.}

\item{\code{scaQThres}}{(scalar)
FDR-corrected p-value cutoff for significance.}

\item{\code{strReport}}{(str)
ImpulseDE2 stdout report.}
}}

\author{
David Sebastian Fischer
}
