% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_classImpulseDE2Object.R
\docType{methods}
\name{ImpulseDE2Object_Accessors}
\alias{ImpulseDE2Object_Accessors}
\alias{get_boolCaseCtrl,ImpulseDE2Object-method}
\alias{get_dfAnnotationProc,ImpulseDE2Object-method}
\alias{get_lsModelFits,ImpulseDE2Object-method}
\alias{get_matCountDataProc,ImpulseDE2Object-method}
\alias{get_scaNProc,ImpulseDE2Object-method}
\alias{get_scaQThres,ImpulseDE2Object-method}
\alias{get_strReport,ImpulseDE2Object-method}
\alias{get_vecConfounders,ImpulseDE2Object-method}
\alias{get_vecDispersions,ImpulseDE2Object-method}
\alias{get_vecSizeFactors,ImpulseDE2Object-method}
\title{ImpulseDE2Object accession methods}
\usage{
\S4method{get_lsModelFits}{ImpulseDE2Object}(object)

\S4method{get_matCountDataProc}{ImpulseDE2Object}(object)

\S4method{get_dfAnnotationProc}{ImpulseDE2Object}(object)

\S4method{get_vecSizeFactors}{ImpulseDE2Object}(object)

\S4method{get_vecDispersions}{ImpulseDE2Object}(object)

\S4method{get_boolCaseCtrl}{ImpulseDE2Object}(object)

\S4method{get_vecConfounders}{ImpulseDE2Object}(object)

\S4method{get_scaNProc}{ImpulseDE2Object}(object)

\S4method{get_scaQThres}{ImpulseDE2Object}(object)

\S4method{get_strReport}{ImpulseDE2Object}(object)
}
\arguments{
\item{object}{(objectImpulseDE2)  A ImpulseDE2 output object.}
}
\value{
The internal data object specified by the function.
}
\description{
Get internal data of ImpulseDE2 output object.
}

