% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_classImpulseDE2Object.R
\docType{methods}
\name{list_accession}
\alias{list_accession}
\alias{names.ImpulseDE2Object}
\alias{names,ImpulseDE2Object-method}
\alias{$,ImpulseDE2Object-method}
\alias{[[,ImpulseDE2Object,character,missing-method}
\alias{list_accession}
\alias{list_accession}
\alias{list_accession}
\title{List-like accessor methods for ImpulseDE2Object}
\usage{
\S4method{names}{ImpulseDE2Object}(x)

\S4method{[[}{ImpulseDE2Object,character,missing}(x, i, j, ...)

\S4method{$}{ImpulseDE2Object}(x, name)
}
\arguments{
\item{x}{(ImpulseDE2Object) ImpulseDE2 output object.}

\item{i, name}{(idx or str) Name or index of core output 
element of ImpulseDE2Object.}

\item{j}{Not used, only vectors.}

\item{...}{Not used.}
}
\value{
Names of core output in ImpulseDE2Object.

Target element from ImpulseDE2Object.

Target element from ImpulseDE2Object.
}
\description{
Allow usage of ImpulseDE2 ouput object like a list with
respect to the core output:
dfImpulseDE2Results and vecDEGenes.
}
\examples{
lsSimulatedData <- simulateDataSetImpulseDE2(
vecTimePointsA   = rep(seq(1,8),3),
vecTimePointsB   = NULL,
vecBatchesA      = NULL,
vecBatchesB      = NULL,
scaNConst        = 30,
scaNImp          = 10,
scaNLin          = 10,
scaNSig          = 10)
objectImpulseDE2 <- runImpulseDE2(
matCountData    = lsSimulatedData$matObservedCounts, 
dfAnnotation    = lsSimulatedData$dfAnnotation,
boolCaseCtrl    = FALSE,
vecConfounders  = NULL,
scaNProc        = 1 )
names(objectImpulseDE2) # Display core output
# With respect to this core output, objectImpulseDE2
# can be treated like a list.
head(objectImpulseDE2[['dfImpulseDE2Results']])
head(objectImpulseDE2$dfImpulseDE2Results)
head(objectImpulseDE2[['vecDEGenes']])
head(objectImpulseDE2$vecDEGenes)

}
\author{
David Sebastian Fischer
}
