% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_classImpulseDE2Object.R
\docType{methods}
\name{names,ImpulseDE2Object-method}
\alias{names,ImpulseDE2Object-method}
\title{List-like accessor methods for ImpulseDE2Object: names()}
\usage{
\S4method{names}{ImpulseDE2Object}(x)
}
\arguments{
\item{x}{(ImpulseDE2Object) ImpulseDE2 output object.}
}
\value{
Names of elements in x available via list-like accessors.
}
\description{
names() function for ImpulseDE2Object.
Allow usage of ImpulseDE2 ouput object like a list with
respect to the most relevant output:
dfImpulseDE2Results and vecDEGenes.
List of all available list-object like accessors:
\link{names,ImpulseDE2Object-method},
\link{[[,ImpulseDE2Object,character,missing-method},
\link{$,ImpulseDE2Object-method}.
}

