% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_fitImpulse.R
\name{optimiseImpulseModelFit}
\alias{optimiseImpulseModelFit}
\title{Fit an impulse model to data of a gene}
\usage{
optimiseImpulseModelFit(vecParamGuess, vecTimepoints, vecCounts,
  scaDispersionEstimate, vecDropoutRate = NULL, vecNormConst,
  vecindTimepointAssign, vecboolObserved, vecboolZero = NULL,
  vecboolNotZeroObserved = NULL, scaWindowRadius = NULL,
  strMode = "batch", MAXIT = 100)
}
\arguments{
\item{vecTimepoints}{(numeric vector number of timepoints) 
Time-points at which gene was sampled.}

\item{vecCounts}{(count vector numer of samples)}

\item{vecindTimepointAssign}{(numeric vector number samples) 
Index of time point assigned to sample in list of sorted
time points (vecX).}

\item{vecParamGuessPeak}{(vector number of parameters [6]) 
Impulse model parameters.}

\item{scaDispersionEstimate:}{(scalar) 
Dispersion estimate for given gene.}

\item{vecDropoutRate:}{(probability vector number of samples) 
Dropout rate estimate for each cell for given gene.}

\item{vecNormConst:}{(numeric vector number of samples) 
Normalisation constants for each sample.}

\item{vecboolObserved:}{(bool vector number of samples)
Whether sample is observed (not NA).}

\item{vecboolZero:}{(bool vector number of samples)
Whether sample has zero count.}

\item{vecboolNotZeroObserved:}{(bool vector number of samples)
Whether sample is not zero and observed (not NA).}

\item{strMode:}{(str) [Default "batch"] 
{"batch","longitudinal","singlecell"}
Mode of model fitting.}

\item{MAXIT:}{(scalar) [Default 100] Number of iterations, which are performed 
to fit the impulse model to the clusters.}
}
\value{
vecFit: (vector [beta, h0, h1, h2, t1, t2, logL_H1, 
   converge_H1) Impulse model parameters, likelihood of
   data under fitted impulse model with given initialiation
   and convergence status of numerical optimisation.
}
\description{
Given a parameter initialisation, this function
performs numerical optimisation using BFGS of the 
likelihood function given the impulse model and returns
the fitted (maximum likelihood) model.
}
\seealso{
Called by \code{fitImpulse_gene}. This function
calls the cost function \code{evalLogLikImpulseBatch_comp},
\code{evalLogLikImpulseTC_comp} or 
\code{evalLogLikImpulseSC_comp} depending on the mode.
}

