% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_plotDEGenes.R
\name{plotDEGenes}
\alias{plotDEGenes}
\title{Plots the impulse fits and data to pdf}
\usage{
plotDEGenes(vecGeneIDs, matCountDataProc, matTranslationFactors = NULL,
  matSizeFactors, dfAnnotationProc, lsImpulseFits, matMuCluster = NULL,
  vecCentroids = NULL, vecClusterAssignments = NULL, dfImpulseResults,
  vecRefPval = NULL, strCaseName, strControlName = NULL,
  strMode = "batch", strSCMode = "clustered", strFileNameSuffix = "",
  strPlotTitleSuffix = "", strPlotSubtitle = "",
  strNameMethod1 = "ImpulseDE2", strNameMethod2 = NULL,
  boolSimplePlot = FALSE, boolLogPlot = FALSE, NPARAM = 6)
}
\arguments{
\item{vecGeneIDs}{(string vector) Gene names to be plotted,}

\item{lsImpulseFits}{(list length 2 or 6) List of matrices which
contain parameter fits and model values for given time course for the
case condition (and control and combined if control is present).
Each parameter matrix is called parameter_'condition' and has the form
(genes x [beta, h0, h1, h2, t1, t2, logL_H1, converge_H1, mu, logL_H0, 
) where beta to t2 are parameters of the impulse
model, mu is the single parameter of the mean model, logL are
log likelihoods of full (H1) and reduced model (H0) respectively, converge
is convergence status of numerical optimisation of model fitting by
\code{optim} from \code{stats}. Each value matrix is called
value_'condition' and has the form (genes x time points) and contains the
counts predicted by the impulse model at the observed time points.}

\item{vecRefPval}{(vec length genes) Method 2 (DESeq2) adjusted p-values}

\item{strCaseName}{(str) Name of the case condition in \code{dfAnnotationProcFull}.}

\item{strFileNameSuffix}{(character string) [Default ""] File extention.}

\item{strPlotSubtitle}{(character string) [Default ""] Subtitle for each plot.}

\item{NPARAM}{(scalar) [Default 6] Number of parameters of impulse model.}

\item{matCountDataProc:}{(matrix genes x samples)
Count data: Reduced version of \code{matCountData}. 
For internal use.}

\item{matTranslationFactors:}{(numeric matrix genes x samples) 
Model scaling factors for each observation which take
longitudinal time series mean within a gene into account 
(translation factors). Computed based based on all samples.}

\item{matSizeFactors:}{(numeric matrix genes x samples) 
Model scaling factors for each observation which take
sequencing depth into account (size factors). One size
factor per sample - rows of this matrix are equal.}

\item{dfAnnotationProc:}{(Table) Processed annotation table. 
Lists co-variables of samples: 
Sample, Condition, Time (numeric), TimeCateg (str)
(and LongitudinalSeries). For internal use.}

\item{dfDEAnalysis}{(data frame genes x fitting characteristics) 
Summary of fitting procedure for each gene.}

\item{strControlName:}{(str) [Default NULL] Name of the control condition in 
\code{dfAnnotationProcFull}.}

\item{strMode:}{(str) [Default "batch"] {"batch","longitudinal","singlecell"}
Mode of model fitting.}

\item{title_string}{(character string) [Default ""] Title for each plot.}

\item{NPARAM}{(scalar) [Default 6] Number of parameters of impulse model.}
}
\description{
Plots the impulse fits and data to pdf.
}
\seealso{
Called by \code{runImpulseDE2}.
}

