% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_plotGenes.R
\name{plotGenes}
\alias{plotGenes}
\title{Plots the impulse fits and data}
\usage{
plotGenes(vecGeneIDs = NULL, scaNTopIDs = NULL, objectImpulseDE2,
  boolCaseCtrl, dirOut = NULL, strFileName = "ImpulseDE2_Trajectories.pdf",
  boolMultiplePlotsPerPage = TRUE, boolSimplePlot = FALSE,
  vecRefPval = NULL, strNameRefMethod = NULL)
}
\arguments{
\item{vecGeneIDs}{(string vector) [Default NULL]
Gene names to be plotted. Must be in rownames of 
objectImpulseDE2@matCountDataProc.
Supply either vecGeneIDs or scaNTopIDs.}

\item{scaNTopIDs}{(int) [Default NULL]
Number of top differentially expressed (by q-value) genes to 
be plotted
Supply either vecGeneIDs or scaNTopIDs.}

\item{objectImpulseDE2}{(ImpulseDE2 object)
Object previously fitted to be used for plotting.}

\item{boolCaseCtrl}{(bool) Whether to create case-ctrl plot.}

\item{dirOut}{(dir) [Default NULL]
Directory into which pdf is printed.}

\item{strFileName}{(str) [Default 'ImpulseDE2_Trajectories.pdf']
File name of pdf with plots.}

\item{boolMultiplePlotsPerPage}{(bool) [Default TRUE]
Whether to create grid with multiple plots on each page of pdf.}

\item{boolSimplePlot}{(bool) [Default TRUE]
Whether to omit batch structure in plotting of model fits
and only plot fit to first batch/all data (if no confounders were given).
This strongly simplifies plots and is recommended e.g. for case-ctrl data.}

\item{vecRefPval}{(vector length vecGeneIDs) [Default NULL]
P/Q-values to be displayed alongside ImpulseDE2 q-value
for differential expression in plot titles.}

\item{strNameRefMethod}{(str) [Default NULL]
Name of reference method used to generate vecRefPval.
Mentioned in plot titles.}
}
\value{
lsgplotsID (gplot list length vecGeneIDs)
List of gplots for IDs in vecGeneIDs. This is secondary 
output next to the .pdf and can be used to extract 
single plots or assemble plots differently.
}
\description{
Plots the impulse fits and data to pdf and return a list of gplots. Points
are size factor normalised data. Consider using boolSimplePlot=TRUE
if the plot seems to crowded.
}
\examples{
lsSimulatedData <- simulateDataSetImpulseDE2(
vecTimePointsA   = rep(seq(1,8),3),
vecTimePointsB   = NULL,
vecBatchesA      = NULL,
vecBatchesB      = NULL,
scaNConst        = 0,
scaNImp          = 40,
scaNLin          = 20,
scaNSig          = 40)
objectImpulseDE2 <- runImpulseDE2(
matCountData    = lsSimulatedData$matObservedCounts, 
dfAnnotation    = lsSimulatedData$dfAnnotation,
boolCaseCtrl    = FALSE,
vecConfounders  = NULL,
boolIdentifyTransients = FALSE,
scaNProc        = 1 )
lsgplotsID <- plotGenes(
scaNTopIDs=5,
objectImpulseDE2=objectImpulseDE2,
boolCaseCtrl=FALSE,
boolMultiplePlotsPerPage=TRUE,
boolSimplePlot=FALSE)
lsgplotsID[[1]]

}
\seealso{
Called by separately by user.
}
\author{
David Sebastian Fischer
}
