% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_processData.R
\name{processData}
\alias{processData}
\title{Process annotation and count data}
\usage{
processData(dfAnnotation = NULL, matCountData = NULL, scaSmallRun = NULL,
  strCaseName = NULL, strControlName = NULL, strMode = NULL,
  strSCMode = NULL, scaWindowRadius = NULL, lsPseudoDE = NULL,
  vecDispersionsExternal = NULL, vecSizeFactorsExternal = NULL,
  boolRunDESeq2 = NULL)
}
\arguments{
\item{strCaseName}{(str) [Default NULL] 
Name of the case condition in \code{dfAnnotation}.}

\item{matCountData:}{(matrix genes x samples) [Default NULL] 
Count data of all conditions, unobserved entries are NA.}

\item{dfAnnotation:}{(Table) [Default NULL] 
Annotation table. Lists co-variables of samples: 
Sample, Condition, Time (and LongitudinalSeries). 
Time must be numeric.}

\item{strControlName:}{(str) [Default NULL] 
Name of the control condition in \code{dfAnnotation}.}

\item{strMode:}{(str) [Default "batch"] 
{"batch","longitudinal","singlecell"}
Mode of model fitting.}
}
\value{
(list length 3) with the following elements:
\itemize{
 \item \code{matCountDataProc}: (count matrix  genes x samples) 
     Count data: Reduced version of \code{matCountData}. 
 \item \code{matProbNB} (probability matrix genes x samples)
     Probability of each observation to originate from the negative
     binomial component in the zero inflated negative binomial
     mixture model. All entries are 1 if not operating in 
     strMode=="singlecell".
}
}
\description{
Check validity of input and process count data matrix and annotation
into data structures used later in \code{runImpulseDE2}.
\code{processData} is structure in the following way:
(I) Subhelper functions:
   checkNull() Check whether object was supplied (is not NULL).
   checkDimMatch() Checks whether dimensions of matrices agree.
   checkElementMatch() Checks whether vectors are identical.
   checkNumeric() Checks whether elements are numeric.
   checkProbability() Checks whether elements are probabilities.
   checkCounts() Checks whether elements are count data.
(II) Helper functions:
   checkData() Check format and presence of input data.
   nameGenes() Name genes if names are not given.
   procAnnotation() Add categorial time variable to annotation table.
   reduceCountData() Reduce count data to data which are utilised later.
(III) Script body
}
\seealso{
Called by \code{runImpulseDE2}.
}

