% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_classImpulseDE2Object.R
\name{set_accessors}
\alias{set_accessors}
\alias{set_boolCaseCtrl}
\alias{set_dfAnnotationProc}
\alias{set_dfImpulseDE2Results}
\alias{set_lsModelFits}
\alias{set_matCountDataProc}
\alias{set_vecAllIDs}
\alias{set_vecConfounders}
\alias{set_vecDEGenes}
\alias{set_vecDispersions}
\alias{set_vecSizeFactors}
\alias{set_scaNProc}
\alias{set_scaQThres}
\alias{set_strReport}
\alias{set_accessors}
\alias{set_boolCaseCtrl}
\alias{set_accessors}
\alias{set_dfAnnotationProc}
\alias{set_accessors}
\alias{set_dfImpulseDE2Results}
\alias{set_accessors}
\alias{set_lsModelFits}
\alias{set_accessors}
\alias{set_matCountDataProc}
\alias{set_accessors}
\alias{set_scaNProc}
\alias{set_accessors}
\alias{set_scaQThres}
\alias{set_accessors}
\alias{set_strReport}
\alias{set_accessors}
\alias{set_vecAllIDs}
\alias{set_accessors}
\alias{set_vecConfounders}
\alias{set_accessors}
\alias{set_vecDEGenes}
\alias{set_accessors}
\alias{set_vecDispersions}
\alias{set_accessors}
\alias{set_vecSizeFactors}
\title{ImpulseDE2Object "set" accession functions}
\usage{
set_boolCaseCtrl(obj, element)

set_dfAnnotationProc(obj, element)

set_dfImpulseDE2Results(obj, element)

set_lsModelFits(obj, element)

set_matCountDataProc(obj, element)

set_scaNProc(obj, element)

set_scaQThres(obj, element)

set_strReport(obj, element)

set_vecAllIDs(obj, element)

set_vecConfounders(obj, element)

set_vecDEGenes(obj, element)

set_vecDispersions(obj, element)

set_vecSizeFactors(obj, element)
}
\arguments{
\item{obj}{(ImpulseDE2Object)  A ImpulseDE2 output object.}

\item{element}{(type depends on element)
An element of the ImpulseDE2 output object which is to be substituted.}
}
\value{
(ImpulseDE2Object)
The ImpulseDE2 object with the target element substituted.
}
\description{
Set internal data of ImpulseDE2 output object.
}
\author{
David Sebastian Fischer
}
