% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_simulateDataSet.R
\name{simulateDataSetImpulseDE2}
\alias{simulateDataSetImpulseDE2}
\title{Simulate a data set for ImpulseDE2}
\usage{
simulateDataSetImpulseDE2(vecTimePointsA, vecTimePointsB, vecBatchesA,
  vecBatchesB, scaNConst, scaNImp, scaNLin, scaNSig, scaNRand = 0,
  scaMumax = 1000, scaSDExpressionChange = 1, scaSDRand = NULL,
  scaMuSizeEffect = 1, scaSDSizeEffect = 0.1, scaMuBatchEffect = NULL,
  scaSDBatchEffect = NULL, dirOutSimulation = NULL)
}
\arguments{
\item{vecTimePointsA}{(numeric vector number of time points)
Number of time points in batch A.}

\item{vecTimePointsB}{(numeric vector number of time points)
Number of time points in batch B.}

\item{vecBatchesA}{(str vector number of samples in vecTimePointsA) [Default NULL]
Batch IDs of each sample in condition A. Set to NULL if 
simulating without batch effects.}

\item{vecBatchesB}{(str vector number of samples in vecTimePointsB) [Default NULL]
Batch IDs of each sample in condition B. Set to NULL if 
simulating without batch effects.}

\item{scaNConst}{(scalar) Number of constant genes in data set.}

\item{scaNImp}{(scalar) Number of impulse distributed genes in data set.}

\item{scaNLin}{(scalar) Number of linear distributed genes in data set.}

\item{scaNSig}{(scalar) Number of sigmoid distributed genes in data set.}

\item{scaNRand}{(scalar) [Default NULL] Number of random distributed genes in data set.}

\item{scaMumax}{(scalar) [Default 1000]
Maximum expression mean parameter to be used.}

\item{scaSDExpressionChange}{(scalar) [Default 1]
Standard deviation of normal distribution from which the 
amplitude change within an impulse trace is drawn.}

\item{scaSDRand}{(scalar) [Default 0]
Standard deviation of normal distribution from which the 
random deviations are drawn.}

\item{scaMuSizeEffect}{(numeric vector number of genes) [Default NULL]
Mean of normal distribution of which scaNLing factor for 
size effects per sample are drawn.}

\item{scaSDSizeEffect}{(numeric vector number of genes) [Default NULL]
Standard deviation of normal distribution of which scaling factor for 
size effects per sample are drawn.}

\item{scaMuBatchEffect}{(numeric vector number of genes) [Default NULL]
Mean of normal distribution of which scaling factor for 
batch effects per gene are drawn (reference is batch A).}

\item{scaSDBatchEffect}{(numeric vector number of genes) [Default NULL]
Standard deviation of normal distribution of which scaling factor for 
batch effects per gene are drawn (reference is batch A).}

\item{dirOutSimulation}{(directory) [Default NULL]
Directory to which simulated parameter objects are 
saved to.}
}
\value{
list (length 2)
\itemize{
   \item dfAnnotation (data frame samples x covariates) 
   {Sample, Condition, Time (numeric), TimeCateg (str)
   (and confounding variables if given).}
   Annotation table with covariates for each sample.
   \item matSampledCountsObserved (matrix genes x cells)
   Sampled count data of all cells after drop-out.
}
}
\description{
Simulates a data set with genes with constant and impulse
expression traces. Expression strength and variation in impulse
like traces are parameterised and random. All temporary files
are saved into dirOutSimulation and only the objects necessary
for running ImpulseDE2 (the count matrix and the annotation table
are returned). The remaining objects representing hidden
parameters can be used to evaluate parameter estimates.
}
\author{
David Sebastian Fischer
}
\seealso{
Called by separately by user.
}

