% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_classImpulseDE2Object.R
\docType{methods}
\name{writeReportToFile,ImpulseDE2Object,character-method}
\alias{writeReportToFile,ImpulseDE2Object,character-method}
\title{Print ImpulseDE2 report string to .txt file}
\usage{
\S4method{writeReportToFile}{ImpulseDE2Object,character}(object, fileReport)
}
\arguments{
\item{object}{(ImpulseDE2Object) Output object of ImpulseDE2.}

\item{fileReport}{(file) File to print report to.}
}
\description{
Print ImpulseDE2 report string to .txt file
}
\author{
David Sebastian Fischer
}

