% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_classImpulseDE2Object.R
\name{writeReportToFile}
\alias{writeReportToFile}
\title{Print ImpulseDE2 report string to .txt file}
\usage{
writeReportToFile(object, fileReport)
}
\arguments{
\item{object}{(ImpulseDE2Object) Output object of ImpulseDE2.}

\item{fileReport}{(file) File to print report to.}
}
\description{
Print ImpulseDE2 report string to .txt file
}

