% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_processData.R
\name{processData}
\alias{processData}
\title{Check and process input to runImpulseDE2()}
\usage{
processData(dfAnnotation, matCountData, boolCaseCtrl, vecConfounders,
  vecDispersionsExternal, vecSizeFactorsExternal)
}
\arguments{
\item{dfAnnotation}{(data frame samples x covariates) 
{Sample, Condition, Time (numeric), TimeCateg (str)
(and confounding variables if given).}
Annotation table with covariates for each sample.}

\item{matCountData}{(matrix genes x samples) [Default NULL] 
Read count data, unobserved entries are NA.}

\item{boolCaseCtrl}{(bool) 
Whether to perform case-control analysis. Does case-only
analysis if FALSE.}

\item{vecConfounders}{(vector of strings number of confounding variables)
Factors to correct for during batch correction. Have to 
supply dispersion factors if more than one is supplied.
Names refer to columns in dfAnnotation.}

\item{vecDispersionsExternal}{(vector length number of
genes in matCountData) [Default NULL]
Externally generated list of gene-wise dispersion factors
which overides DESeq2 generated dispersion factors.}

\item{vecSizeFactorsExternal}{(vector length number of
cells in matCountData) [Default NULL]
Externally generated list of size factors which override
size factor computation in ImpulseDE2.}
}
\value{
(list length 4)
\itemize{
   \item matCountDataProc (matrix genes x samples)
   Read count data.
   \item dfAnnotationProc (data frame samples x covariates) 
   {Sample, Condition, Time (numeric), TimeCateg (str)
   (and confounding variables if given).}
   Processed annotation table with covariates for each sample.
   \item vecSizeFactorsExternalProc (numeric vector number of samples) 
   Model scaling factors for each sample which take
   sequencing depth into account (size factors).
   \item vecDispersionsExternalProc (vector number of genes) Gene-wise 
   negative binomial dispersion hyper-parameter.
   \item strReportProcessing (str) String of stdout of processData().
}
}
\description{
Check validity of input and process count data matrix and annotation
into data structures used later in \link{runImpulseDE2}.
\link{processData} is structure in the following way:
\itemize{
   \item Subhelper functions:
   \itemize{
     \item checkNull() Check whether object was supplied (is not NULL).
     \item checkDimMatch() Checks whether dimensions of matrices agree.
     \item checkElementMatch() Checks whether vectors are identical.
     \item checkNumeric() Checks whether elements are numeric.
     \item checkProbability() Checks whether elements are probabilities.
     \item checkCounts() Checks whether elements are count data.
   }
   \item Helper functions:
   \itemize{
     \item checkData() Check format and presence of input data.
     \item nameGenes() Name genes if names are not given.
     \item procAnnotation() Add categorial time variable to 
     annotation table.
     Add nested batch column if necessary.
     Reduce to samples used.
     \item reduceCountData() Reduce count data to data which are 
     utilised later.
   }
   \item Script body
}
}
\seealso{
Called by \link{runImpulseDE2}.
}
\author{
David Sebastian Fischer
}
