% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_runDESeq2.R
\name{runDESeq2}
\alias{runDESeq2}
\title{Wrapper function for running DESeq2}
\usage{
runDESeq2(dfAnnotationProc, matCountDataProc, nProc = 1, strCaseName = NULL,
  strControlName = NULL, strMode = "batch")
}
\arguments{
\item{matCountDataProc:}{(matrix genes x samples)
Count data: Reduced version of \code{matCountData}. 
For internal use.}

\item{dfAnnotationProc:}{(Table) Processed annotation table. 
Lists co-variables of samples: 
Sample, Condition, Time (numeric), TimeCateg (str)
(and Timecourse). For internal use.}

\item{nProc:}{(scalar) [Default 1] 
Number of processes for parallelisation.}

\item{strCaseName:}{(str) [Default NULL] 
Name of the case condition in \code{dfAnnotation}.}

\item{strControlName:}{(str) [Default NULL] 
Name of the control condition in \code{dfAnnotation}.}

\item{strMode:}{(str) [Default "batch"] 
{"batch","longitudinal","singlecell"}
Mode of model fitting.}
}
\value{
(list length 2) with the following elements:
\itemize{
 \item \code{dds_dispersions} (vector number of genes) Inverse of gene-wise 
     negative binomial dispersion coefficients computed by DESeq2.
 \item \code{dds_resultsTable} (data frame) DESeq2 results.
}
}
\description{
Run DESeq2 and extract differential expression analysis results and 
overdispersion coefficients.
}
\seealso{
Called by \code{runImpulseDE2}.
}

