% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srcImpulseDE2_classImpulseDE2Object.R
\name{writeReportToFile}
\alias{writeReportToFile}
\title{Print ImpulseDE2 report to .txt file}
\usage{
writeReportToFile(object, fileReport)
}
\arguments{
\item{object}{(ImpulseDE2Object) Output object of ImpulseDE2.}

\item{fileReport}{(file) File to print report to.}
}
\value{
No return.
}
\description{
Print ImpulseDE2 report to .txt file.
}
\examples{
dirPWD <- getwd() # Will save into current working directory.
lsSimulatedData <- simulateDataSetImpulseDE2(
vecTimePointsA   = rep(seq(1,8),3),
vecTimePointsB   = NULL,
vecBatchesA      = NULL,
vecBatchesB      = NULL,
scaNConst        = 30,
scaNImp          = 10,
scaNLin          = 10,
scaNSig          = 10)
objectImpulseDE2 <- runImpulseDE2(
matCountData    = lsSimulatedData$matObservedCounts, 
dfAnnotation    = lsSimulatedData$dfAnnotation,
boolCaseCtrl    = FALSE,
vecConfounders  = NULL,
scaNProc        = 1 )
# Uncomment to run:
#writeReportToFile(
#object=objectImpulseDE2,
#file=paste0(dirPWD, 'ImpulseDE2Report.txt')
#)

}
\author{
David Sebastian Fischer
}
